requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-asyncio==0.23.0
anthropic==0.76.0
Jinja2==3.1.3
